requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
discord==2.3.2
discord.py==2.3.2
exceptiongroup==1.2.1
execnet==2.1.1
fasteners==0.19
filelock==3.14.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
pdbp==1.5.0
platformdirs==4.2.1
pluggy==1.5.0
py==1.11.0
Pygments==2.17.2
pynose==1.5.1
pyotp==2.9.0
PySocks==1.7.1
pytest==8.2.0
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.3
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
sbvirtualdisplay==1.3.0
selenium==4.20.0
seleniumbase==4.26.3
six==1.16.0
slugify==0.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tabcompleter==1.3.0
text-unidecode==1.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
